+++ src/Lakona.Tool/Scaffolding/Templates/ChatClientTemplates.cs
@@ -633,6 +633,7 @@ internal static class ChatClientTemplates
         using Godot;
         using Shared.Contracts.Chat;
         using Client.Chat;
+        using Client.Login;
 
         namespace Client.Chat
         {
@@ -903,6 +904,7 @@ internal static class ChatClientTemplates
         using System.Threading.Tasks;
         using Shared.Contracts.Chat;
         using Client.Chat;
+        using Client.Login;
         using UnityEngine;
         using UnityEngine.UIElements;